+++ app/src/main/java/xyz/goodistory/yumemiassignment/GitHubService.kt
@@ -11,27 +11,7 @@ interface GitHubService {
     @GET("users/{login}")
     fun getUser(@Path("login") login: String): Call<User>
 
-    data class Contributor(
-        val login: String,
-        val id: Int,
-        val node_id: String,
-        val avatar_url: String,
-        val gravatar_id: String,
-        val url: String,
-        val html_url: String,
-        val followers_url: String,
-        val following_url: String,
-        val gists_url: String,
-        val starred_url: String,
-        val subscriptions_url: String,
-        val organizations_url: String,
-        val repos_url: String,
-        val events_url: String,
-        val received_events_url: String,
-        val type: String,
-        val site_admin: Boolean,
-        val contributions: Int,
-        )
+
 
     data class User(
         val login: String,
