+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <resources>
     <string name="app_name">YumemiAssignment</string>
     <string name="api_endpoint">https://api.github.com/</string>
+    <string name="contributors_list_item_contributions_label">Contribute数: </string>
     <string name="contributors_list_item_detail">詳細</string>
 </resources>
